# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for tracking and managing shisha (hookah) sessions",
        "title": "Shisha Log API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/toof-jp/shisha-log"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/creators/stats": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get creator statistics for the authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "statistics"
                ],
                "summary": "Get creator statistics",
                "responses": {
                    "200": {
                        "description": "Creator statistics",
                        "schema": {
                            "$ref": "#/definitions/models.CreatorStats"
                        }
                    },
                    "500": {
                        "description": "Failed to get creator statistics",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/flavors/stats": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get flavor usage statistics for the authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "statistics"
                ],
                "summary": "Get flavor statistics",
                "responses": {
                    "200": {
                        "description": "Flavor statistics",
                        "schema": {
                            "$ref": "#/definitions/models.FlavorStats"
                        }
                    },
                    "500": {
                        "description": "Failed to get flavor statistics",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/sessions": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get paginated list of sessions for the authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sessions"
                ],
                "summary": "Get user's sessions",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Number of items per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Number of items to skip",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated sessions list",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "limit": {
                                    "type": "integer"
                                },
                                "offset": {
                                    "type": "integer"
                                },
                                "sessions": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/models.SessionWithFlavors"
                                    }
                                },
                                "total": {
                                    "type": "integer"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to get sessions",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Create a new shisha session for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sessions"
                ],
                "summary": "Create a new session",
                "parameters": [
                    {
                        "description": "Session data",
                        "name": "session",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateSessionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created session with flavors",
                        "schema": {
                            "$ref": "#/definitions/models.SessionWithFlavors"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to create session",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/sessions/by-date": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get all sessions for a specific date",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sessions"
                ],
                "summary": "Get sessions by date",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Date in YYYY-MM-DD format",
                        "name": "date",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Timezone (default UTC)",
                        "name": "timezone",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sessions for the specified date",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "date": {
                                    "type": "string"
                                },
                                "sessions": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/models.SessionWithFlavors"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid date format",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to get sessions",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/sessions/calendar": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get session counts for each day in a specific month",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sessions"
                ],
                "summary": "Get calendar data",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Year",
                        "name": "year",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Month (1-12)",
                        "name": "month",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Timezone (default UTC)",
                        "name": "timezone",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Map of date strings to session counts",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "integer"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid parameters",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to get calendar data",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/stores/stats": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get store visit statistics for the authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "statistics"
                ],
                "summary": "Get store statistics",
                "responses": {
                    "200": {
                        "description": "Store statistics",
                        "schema": {
                            "$ref": "#/definitions/models.StoreStats"
                        }
                    },
                    "500": {
                        "description": "Failed to get store statistics",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.CreateFlavorRequest": {
            "type": "object",
            "properties": {
                "brand": {
                    "type": "string"
                },
                "flavor_name": {
                    "type": "string"
                }
            }
        },
        "models.CreateSessionRequest": {
            "type": "object",
            "required": [
                "session_date"
            ],
            "properties": {
                "creator": {
                    "type": "string"
                },
                "flavors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.CreateFlavorRequest"
                    }
                },
                "mix_name": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "order_details": {
                    "type": "string"
                },
                "session_date": {
                    "type": "string"
                },
                "store_name": {
                    "type": "string"
                }
            }
        },
        "models.CreatorCount": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "creator": {
                    "type": "string"
                }
            }
        },
        "models.CreatorStats": {
            "type": "object",
            "properties": {
                "creators": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.CreatorCount"
                    }
                }
            }
        },
        "models.FlavorCount": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "flavor_name": {
                    "type": "string"
                }
            }
        },
        "models.FlavorStats": {
            "type": "object",
            "properties": {
                "all_flavors": {
                    "description": "All flavors",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.FlavorCount"
                    }
                },
                "main_flavors": {
                    "description": "First flavors only",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.FlavorCount"
                    }
                }
            }
        },
        "models.SessionFlavor": {
            "type": "object",
            "properties": {
                "brand": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "flavor_name": {
                    "type": "string"
                },
                "flavor_order": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "session_id": {
                    "type": "string"
                }
            }
        },
        "models.SessionWithFlavors": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "creator": {
                    "type": "string"
                },
                "flavors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SessionFlavor"
                    }
                },
                "id": {
                    "type": "string"
                },
                "mix_name": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "order_details": {
                    "type": "string"
                },
                "session_date": {
                    "type": "string"
                },
                "store_name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "models.StoreCount": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "store_name": {
                    "type": "string"
                }
            }
        },
        "models.StoreStats": {
            "type": "object",
            "properties": {
                "stores": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.StoreCount"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}